+++ src/v1/apis/auth/auth.service.ts
@@ -3,14 +3,14 @@ import { FastifyBaseLogger } from 'fastify';
 import { JWT } from '@fastify/jwt';
 import { v4 as uuidv4 } from 'uuid';
 
-import prisma from '../utils/prisma.js';
+import prisma from '../..//common/utils/prisma.js';
 import {
   loginRequestSchema,
   loginResponseSchema,
   signupRequestSchema,
   signupResponseSchema,
-} from '../schemas/auth.schema.js';
-import { STATUS } from '../constants/status.js';
+} from '../..//common/schemas/auth.schema.js';
+import { STATUS } from '../..//common/constants/status.js';
 
 export async function signupService(
   data: z.infer<typeof signupRequestSchema>,
