+++ src/v1/common/exceptions/core.error.ts
