+++ src/v1/common/schemas/core.schema.ts
