+++ src/v1/common/types/fastify.d.ts
