+++ src/v1/common/types/global.d.ts
