+++ src/v1/index.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance } from 'fastify';
 
-import authRoutes from './routes/auth.route.js';
-import usersRoutes from './routes/users.route.js';
+import authRoutes from './apis/auth/auth.route.js';
+import usersRoutes from './apis/users/users.route.js';
 
 export default async function routeV1(fastify: FastifyInstance) {
   fastify.register(authRoutes, { prefix: '/auth' });
